requirements.txt
faust #==1.10.4
pymongo==3.10.1
#kafka-python==2.0.1
kafka-python==2.0.2
